# docs/api/openapi.json
{"openapi":"3.1.0","info":{"title":"API Humedal La Orquídea","description":"Backend de la Plataforma de Educación Ambiental e Interactividad del Humedal La Orquídea","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"jwt":[]}],"tags":[{"name":"Autenticación","description":"Inicio de sesión y datos de la sesión actual (HU-1)"},{"name":"Administradores","description":"Gestión de cuentas de administrador, solo SUPERADMINISTRADOR (HU-5)"},{"name":"Fichas taxonómicas","description":"Fichas de especies del humedal (HU-7, HU-8)"}],"paths":{"/api/fichas-taxonomicas/{id}":{"get":{"tags":["Fichas taxonómicas"],"summary":"Consultar ficha taxonómica","description":"Público.","operationId":"obtenerPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Ficha encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaxonDto"}}}},"404":{"description":"La ficha no existe","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}},"security":[]},"put":{"tags":["Fichas taxonómicas"],"summary":"Editar ficha taxonómica","description":"Solo ADMINISTRADOR o SUPERADMINISTRADOR. Reemplaza todos los datos de la ficha con el mismo formulario de la creación; la foto es opcional y, si no se envía, se conserva la actual. Ante un 200 el cliente muestra: 'La ficha se actualizó correctamente.'","operationId":"actualizar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/TaxonRequest"}}}},"responses":{"200":{"description":"Ficha actualizada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaxonDto"}}}},"400":{"description":"Debes completar todos los campos obligatorios. / El formato de la imagen no es válido.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"401":{"description":"Sin sesión iniciada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"403":{"description":"El usuario no es administrador","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"404":{"description":"La ficha no existe","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"409":{"description":"Otra ficha ya usa ese nombre científico","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"413":{"description":"La imagen supera el tamaño máximo","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}}}},"/api/administradores/{id}":{"put":{"tags":["Administradores"],"summary":"Editar o inhabilitar un administrador","description":"Reemplaza nombre, apellido, correo, teléfono y estado (habilitado) de la cuenta. Ante un 200 el cliente muestra: 'Los cambios se guardaron correctamente.'","operationId":"actualizar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdministratorUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Administrador actualizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdministratorDto"}}}},"400":{"description":"Debes completar todos los campos obligatorios. / Ingresa un correo electrónico válido.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"401":{"description":"Sin sesión iniciada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"403":{"description":"El usuario no es superadministrador","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"404":{"description":"El administrador no existe","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"409":{"description":"Este correo ya está registrado.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/fichas-taxonomicas":{"get":{"tags":["Fichas taxonómicas"],"summary":"Listar fichas taxonómicas","description":"Público. Ordenadas por nombre común. Con el parámetro categoria (AVE, PLANTA o INSECTO) devuelve solo las de ese listado; sin él, todas. Si la categoría no tiene fichas responde una lista vacía y el cliente muestra: 'Aún no hay especies registradas en esta categoría.'","operationId":"listar","parameters":[{"name":"categoria","in":"query","description":"Categoría a consultar; si se omite, se listan todas","required":false,"schema":{"type":"string","enum":["AVE","PLANTA","INSECTO"]},"example":"AVE"}],"responses":{"200":{"description":"Listado de fichas (vacío si no hay ninguna)","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaxonDto"}}}}},"400":{"description":"El valor del parámetro 'categoria' no es válido.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}},"security":[]},"post":{"tags":["Fichas taxonómicas"],"summary":"Crear ficha taxonómica","description":"Solo ADMINISTRADOR o SUPERADMINISTRADOR. Formulario multipart con los datos y la foto (jpg o png, máx. 10 MB).","operationId":"crear","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/TaxonRequest"}}}},"responses":{"201":{"description":"Ficha creada y publicada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaxonDto"}}}},"400":{"description":"Debes completar todos los campos obligatorios. / El formato de la imagen no es válido.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"401":{"description":"Sin sesión iniciada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"403":{"description":"El usuario no es administrador","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"409":{"description":"Ya existe una ficha con ese nombre científico","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"413":{"description":"La imagen supera el tamaño máximo","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}}}},"/api/autenticacion/restablecer-contrasena":{"post":{"tags":["Autenticación"],"summary":"Restablecer contraseña","description":"Cambia la contraseña con el código recibido por correo. El código es de un solo uso y se invalida al vencer o tras varios intentos fallidos.","operationId":"restablecerContrasena","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetRequest"}}},"required":true},"responses":{"204":{"description":"Contraseña actualizada"},"400":{"description":"Datos inválidos, o código incorrecto o vencido.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}},"security":[]}},"/api/autenticacion/registro":{"post":{"tags":["Autenticación"],"summary":"Registrar nuevo usuario","description":"Registra un nuevo usuario en el sistema y asigna el rol de usuario registrado.","operationId":"registrarUsuario","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Registro exitoso creado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}},"400":{"description":"Los datos enviados no son válidos.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"409":{"description":"El correo electrónico ya está registrado.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}},"security":[]}},"/api/autenticacion/recuperar-contrasena":{"post":{"tags":["Autenticación"],"summary":"Solicitar código de recuperación","description":"Envía un código de 6 dígitos al correo si pertenece a una cuenta habilitada. Responde igual aunque el correo no esté registrado, para no revelar qué cuentas existen.","operationId":"solicitarCodigoRecuperacion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordRecoveryRequest"}}},"required":true},"responses":{"204":{"description":"Solicitud recibida"},"400":{"description":"El correo es obligatorio y debe ser válido.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}},"security":[]}},"/api/autenticacion/iniciar-sesion":{"post":{"tags":["Autenticación"],"summary":"Iniciar sesión","description":"Valida correo y contraseña y devuelve un JWT junto con el rol del usuario.","operationId":"iniciarSesion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Ingreso exitoso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"400":{"description":"Ambos campos son obligatorios.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"401":{"description":"Correo o contraseña incorrectos.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}},"security":[]}},"/api/administradores/{id}/revocar-acceso":{"post":{"tags":["Administradores"],"summary":"Revocar el acceso de administrador","description":"Degrada la cuenta a USUARIO_REGISTRADO y notifica al afectado por correo.","operationId":"revocarAcceso","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Acceso revocado; la cuenta quedó como usuario registrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdministratorDto"}}}},"401":{"description":"Sin sesión iniciada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"403":{"description":"El usuario no es superadministrador","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"404":{"description":"El administrador no existe","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"409":{"description":"No puedes revocar el único superadministrador de la plataforma.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/administradores/registrarAdmin":{"post":{"tags":["Administradores"],"summary":"Registrar un nuevo administrador","description":"Crea una nueva cuenta de administrador con los datos proporcionados.","operationId":"registrar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Administrador registrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}},"400":{"description":"Debes completar todos los campos obligatorios. / Ingresa un correo electrónico válido.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"401":{"description":"Sin sesión iniciada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"403":{"description":"El usuario no es superadministrador","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}},"409":{"description":"Este correo ya está registrado.","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/autenticacion/yo":{"get":{"tags":["Autenticación"],"summary":"Usuario de la sesión actual","description":"Devuelve los datos y el rol del dueño del token; útil para restaurar la sesión en el cliente.","operationId":"obtenerUsuarioActual","responses":{"200":{"description":"Usuario autenticado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticatedUserDto"}}}},"401":{"description":"Token ausente, inválido o vencido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}}}}},"components":{"schemas":{"TaxonRequest":{"type":"object","description":"Datos para crear una ficha taxonómica","properties":{"categoria":{"type":"string","description":"Listado en el que se publica la ficha","enum":["AVE","PLANTA","INSECTO"],"example":"AVE"},"orden":{"type":"string","description":"Orden (DwC: order)","example":"Passeriformes","maxLength":100,"minLength":0},"familia":{"type":"string","description":"Familia (DwC: family)","example":"Thraupidae","maxLength":100,"minLength":0},"genero":{"type":"string","description":"Género (DwC: genus)","example":"Thraupis","maxLength":100,"minLength":0},"nombreCientifico":{"type":"string","description":"Nombre científico (DwC: scientificName); único en la plataforma","example":"Thraupis episcopus","maxLength":200,"minLength":0},"nombreComun":{"type":"string","description":"Nombre común (DwC: vernacularName)","example":"Azulejo","maxLength":150,"minLength":0},"alimentacion":{"type":"string","description":"Alimentación de la especie","example":"Frutos, néctar e insectos pequeños","maxLength":2000,"minLength":0},"rolEnHumedal":{"type":"string","description":"Rol ecológico de la especie en el humedal","example":"Dispersor de semillas","maxLength":2000,"minLength":0},"estadoConservacion":{"type":"string","description":"Categoría UICN","enum":["EX","EW","CR","EN","VU","NT","LC","DD","NE"],"example":"LC"},"foto":{"type":"string","format":"binary","description":"Foto de la especie, jpg o png, máximo 10 MB. Obligatoria al crear; al editar, si no se envía (o llega vacía) se conserva la actual"}},"required":["alimentacion","categoria","estadoConservacion","familia","genero","nombreCientifico","nombreComun","orden","rolEnHumedal"]},"ApiErrorResponse":{"type":"object","description":"Formato uniforme de las respuestas de error","properties":{"fecha":{"type":"string","format":"date-time","description":"Momento del error","example":"2026-09-28T15:30:00Z"},"estado":{"type":"integer","format":"int32","description":"Código HTTP","example":401},"error":{"type":"string","description":"Nombre del código HTTP","example":"Unauthorized"},"mensaje":{"type":"string","description":"Mensaje para mostrar al usuario","example":"Correo o contraseña incorrectos."},"ruta":{"type":"string","description":"Ruta solicitada","example":"/api/autenticacion/iniciar-sesion"}}},"TaxonDto":{"type":"object","description":"Ficha taxonómica publicada","properties":{"id":{"type":"string","format":"uuid","description":"Identificador de la ficha","example":"5b0c7f0e-2f5d-4a8e-9a57-0c7d3c1b6e21"},"categoria":{"type":"string","description":"Listado en el que se publica la ficha","enum":["AVE","PLANTA","INSECTO"],"example":"AVE"},"orden":{"type":"string","description":"Orden (DwC: order)","example":"Passeriformes"},"familia":{"type":"string","description":"Familia (DwC: family)","example":"Thraupidae"},"genero":{"type":"string","description":"Género (DwC: genus)","example":"Thraupis"},"nombreCientifico":{"type":"string","description":"Nombre científico (DwC: scientificName)","example":"Thraupis episcopus"},"nombreComun":{"type":"string","description":"Nombre común (DwC: vernacularName)","example":"Azulejo"},"alimentacion":{"type":"string","description":"Alimentación de la especie","example":"Frutos, néctar e insectos pequeños"},"rolEnHumedal":{"type":"string","description":"Rol ecológico de la especie en el humedal","example":"Dispersor de semillas"},"estadoConservacion":{"type":"string","description":"Categoría UICN","enum":["EX","EW","CR","EN","VU","NT","LC","DD","NE"],"example":"LC"},"urlFoto":{"type":"string","description":"URL pública de la foto","example":"http://localhost:9000/orquidea/fichas/5b0c7f0e.jpg"},"fechaCreacion":{"type":"string","format":"date-time","description":"Fecha de creación","example":"2026-09-28T15:30:00Z"},"fechaActualizacion":{"type":"string","format":"date-time","description":"Fecha de la última modificación","example":"2026-09-28T15:30:00Z"}}},"AdministratorUpdateRequest":{"type":"object","description":"Datos editables de un administrador","properties":{"nombre":{"type":"string","description":"Nombre","example":"María","maxLength":100,"minLength":0},"apellido":{"type":"string","description":"Apellido","example":"Pérez","maxLength":100,"minLength":0},"correo":{"type":"string","format":"email","description":"Correo; único en la plataforma","example":"admin@orquidea.local","maxLength":254,"minLength":0},"habilitado":{"type":"boolean","description":"false inhabilita la cuenta: no podrá iniciar sesión","example":true}},"required":["apellido","correo","habilitado","nombre"]},"ApiResponse":{},"AdministratorDto":{"type":"object","description":"Administrador de la plataforma","properties":{"id":{"type":"string","format":"uuid","description":"Identificador del usuario","example":"3f6c1a52-7a1e-4c5b-9c1d-2b8f0e4a9d10"},"nombre":{"type":"string","description":"Nombre","example":"María"},"apellido":{"type":"string","description":"Apellido","example":"Pérez"},"correo":{"type":"string","description":"Correo","example":"admin@orquidea.local"},"rol":{"type":"string","description":"Rol actual; tras revocar el acceso pasa a USUARIO_REGISTRADO","enum":["SUPERADMINISTRADOR","ADMINISTRADOR","USUARIO_REGISTRADO"],"example":"ADMINISTRADOR"},"habilitado":{"type":"boolean","description":"Si la cuenta puede iniciar sesión","example":true},"fechaCreacion":{"type":"string","format":"date-time","description":"Fecha de creación de la cuenta","example":"2026-09-28T15:30:00Z"}}},"PasswordResetRequest":{"type":"object","description":"Código recibido por correo y nueva contraseña","properties":{"correo":{"type":"string","description":"Correo de la cuenta a recuperar","example":"admin@orquidea.local","minLength":1},"codigo":{"type":"string","description":"Código de 6 dígitos enviado al correo","example":"482913","minLength":1,"pattern":"\\d{6}"},"contrasena":{"type":"string","description":"Nueva contraseña","example":"MiClaveNueva123","minLength":1}},"required":["codigo","contrasena","correo"]},"RegisterRequest":{"type":"object","description":"Datos necesarios para registrar un nuevo usuario.","properties":{"nombre":{"type":"string","description":"Nombre del usuario.","example":"Santiago","maxLength":2147483647,"minLength":2},"apellido":{"type":"string","description":"Apellido del usuario.","example":"Bastidas","maxLength":2147483647,"minLength":2},"correo":{"type":"string","format":"email","description":"Correo electrónico que identificará al usuario y será utilizado para iniciar sesión.","example":"santiago.bastidas@gmail.com","minLength":1},"contrasena":{"type":"string","description":"Contraseña que utilizará el usuario para iniciar sesión.","example":"123456","minLength":1}},"required":["apellido","contrasena","correo","nombre"]},"RegisterResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"nombre":{"type":"string"},"apellido":{"type":"string"},"correo":{"type":"string"},"rol":{"type":"string","enum":["SUPERADMINISTRADOR","ADMINISTRADOR","USUARIO_REGISTRADO"]}}},"PasswordRecoveryRequest":{"type":"object","description":"Solicitud de un código para restablecer la contraseña","properties":{"correo":{"type":"string","format":"email","description":"Correo de la cuenta a recuperar","example":"admin@orquidea.local","minLength":1}},"required":["correo"]},"LoginRequest":{"type":"object","description":"Credenciales para iniciar sesión","properties":{"correo":{"type":"string","description":"Correo del usuario","example":"admin@orquidea.local","minLength":1},"contrasena":{"type":"string","description":"Contraseña del usuario","example":"MiClaveSegura123","minLength":1}},"required":["contrasena","correo"]},"AuthenticatedUserDto":{"type":"object","description":"Datos básicos del usuario con sesión iniciada","properties":{"id":{"type":"string","format":"uuid","description":"Identificador del usuario","example":"3f6c1a52-7a1e-4c5b-9c1d-2b8f0e4a9d10"},"nombre":{"type":"string","description":"Nombre","example":"María"},"apellido":{"type":"string","description":"Apellido","example":"Pérez"},"correo":{"type":"string","description":"Correo","example":"admin@orquidea.local"},"rol":{"type":"string","description":"Rol del usuario; define las opciones habilitadas en la aplicación","enum":["SUPERADMINISTRADOR","ADMINISTRADOR","USUARIO_REGISTRADO"],"example":"ADMINISTRADOR"}}},"LoginResponse":{"type":"object","description":"Resultado de un inicio de sesión exitoso","properties":{"token":{"type":"string","description":"JWT a enviar en el encabezado Authorization: Bearer <token>"},"tipo":{"type":"string","description":"Tipo de token","example":"Bearer"},"expiraEnSegundos":{"type":"integer","format":"int64","description":"Segundos de validez del token","example":28800},"usuario":{"$ref":"#/components/schemas/AuthenticatedUserDto","description":"Usuario autenticado"}}}},"securitySchemes":{"jwt":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}